+++ src/pages/MyPage/MyPage.jsx
@@ -9,7 +9,7 @@
 import {Outlet } from 'react-router-dom';
 import SettingsNavbar from "./SettingsNavbar.jsx";
 import useTitle from "../../components/useTitle.jsx";
-import CheckLogin from "../../components/CheckLogin.jsx";
+import CheckLogin from "../../static/js/checkLogin.js";
 import AccessDeniedPage from "../AccessDeniedPage/AccessDeniedPage.jsx";
 import Logout from "../../components/Logout.jsx";
 import "./MyPage.css";
