+++ src/pages/SettingsPage/SettingsPage.jsx
@@ -1,4 +1,3 @@
-import ProviderSettings from "../../components/ProviderSettings.jsx";
 import {Outlet} from "react-router-dom";
 
 //TODO: Link to manage cars. Change the routing so the buttons dont persist to pages it shouldn't
