+++ src/static/js/checkLogin.js
